+++ index.js
@@ -1,5 +1,4 @@
 'use strict';
-const semver = require('semver');
 const execa = require('execa');
 const del = require('del');
 const Listr = require('listr');
@@ -8,9 +7,8 @@ require('any-observable/register/rxjs-all');
 const Observable = require('any-observable');
 const streamToObservable = require('stream-to-observable');
 const readPkgUp = require('read-pkg-up');
-
-const VERSIONS = ['major', 'minor', 'patch', 'premajor', 'preminor', 'prepatch', 'prerelease'];
-const PRERELEASE_VERSIONS = ['premajor', 'preminor', 'prepatch', 'prerelease'];
+const prerequisiteTasks = require('./lib/prerequisite');
+const gitTasks = require('./lib/git');
 
 const exec = (cmd, args) => {
 	// Use `Observable` support if merged https://github.com/sindresorhus/execa/pull/26
@@ -20,98 +18,6 @@ const exec = (cmd, args) => {
 		streamToObservable(cp.stdout.pipe(split()), {await: cp}),
 		streamToObservable(cp.stderr.pipe(split()), {await: cp})
 	).filter(Boolean);
-};
-
-const prerequisiteCheckTasks = (input, pkg, opts) => {
-	const newVersion = VERSIONS.indexOf(input) === -1 ? input : semver.inc(pkg.version, input);
-
-	const tasks = [
-		{
-			title: 'Validate version',
-			task: () => {
-				if (VERSIONS.indexOf(input) === -1 && !semver.valid(input)) {
-					return Promise.reject(new Error(`Version should be either ${VERSIONS.join(', ')}, or a valid semver version.`));
-				}
-
-				if (semver.gte(pkg.version, newVersion)) {
-					return Promise.reject(new Error(`New version \`${newVersion}\` should be higher than current version \`${pkg.version}\``));
-				}
-			}
-		},
-		{
-			title: 'Check for pre-release version',
-			task: () => {
-				if ((PRERELEASE_VERSIONS.indexOf(input) !== -1 || semver.prerelease(input)) && !opts.tag) {
-					return Promise.reject(new Error('You must specify a dist-tag using --tag when publishing a pre-release version. This prevents accidentally tagging unstable versions as "latest". https://docs.npmjs.com/cli/dist-tag'));
-				}
-			}
-		},
-		{
-			title: 'Check npm version',
-			task: () => execa.stdout('npm', ['version', '--json']).then(json => {
-				const versions = JSON.parse(json);
-				if (semver.gte(process.version, '6.0.0') && !semver.satisfies(versions.npm, '>=2.15.8 <3.0.0 || >=3.10.1')) {
-					return Promise.reject(new Error(`npm@${versions.npm} has known issues publishing when running Node.js 6. Please upgrade npm or downgrade Node and publish again. https://github.com/npm/npm/issues/5082`));
-				}
-			})
-		},
-		{
-			title: 'Check git tag existence',
-			task: () => execa('git', ['fetch'])
-				.then(() => execa.stdout('git', ['rev-parse', '--quiet', '--verify', `refs/tags/v${newVersion}`]))
-				.then(
-					output => {
-						if (output) {
-							throw new Error(`Git tag \`v${newVersion}\` already exists.`);
-						}
-					},
-					err => {
-						// Command fails with code 1 and no output if the tag does not exist, even though `--quiet` is provided
-						// https://github.com/sindresorhus/np/pull/73#discussion_r72385685
-						if (err.stdout !== '' || err.stderr !== '') {
-							throw err;
-						}
-					}
-				)
-		}
-	];
-
-	return new Listr(tasks);
-};
-
-const gitTasks = opts => {
-	const tasks = [
-		{
-			title: 'Check current branch',
-			task: () => execa.stdout('git', ['symbolic-ref', '--short', 'HEAD']).then(branch => {
-				if (branch !== 'master') {
-					throw new Error('Not on `master` branch. Use --any-branch to publish anyway.');
-				}
-			})
-		},
-		{
-			title: 'Check local working tree',
-			task: () => execa.stdout('git', ['status', '--porcelain']).then(status => {
-				if (status !== '') {
-					throw new Error('Unclean working tree. Commit or stash changes first.');
-				}
-			})
-		},
-		{
-			title: 'Check remote history',
-			task: () => execa.stdout('git', ['rev-list', '--count', '--left-only', '@{u}...HEAD']).then(result => {
-				if (result !== '0') {
-					throw new Error('Remote history differ. Please pull changes.');
-				}
-			})
-		}
-	];
-
-	if (opts.anyBranch) {
-		tasks.shift();
-	}
-
-	return new Listr(tasks);
 };
 
 module.exports = (input, opts) => {
@@ -125,7 +31,7 @@ module.exports = (input, opts) => {
 	const tasks = new Listr([
 		{
 			title: 'Prerequisite check',
-			task: () => prerequisiteCheckTasks(input, pkg, opts)
+			task: () => prerequisiteTasks(input, pkg, opts)
 		},
 		{
 			title: 'Git',
